[ALPS02718870] Audio: Refine log
[Detail]
Refine log and fix log build error.

MTK-Commit-Id: 2e5317515d4988b35841c7df507a2658167f5aab

CR-Id: ALPS02718870
Feature: Audio Codec

+++ sound/soc/mediatek/common/mt_soc_pcm_dl2.c
@@ -524,8 +524,8 @@ void mtk_dl2_copy2buffer(const void *addr, uint32_t size)
 {
 	bool again = false;
 
-	PRINTK_AUD_DL2("%s, addr 0x%p 0x%p, size %d %d\n", __func__, (int)addr,
-			(int)ISRCopyBuffer.pBufferBase, size, ISRCopyBuffer.u4BufferSize);
+	PRINTK_AUD_DL2("%s, addr 0x%p 0x%p, size %d %d\n", __func__, addr,
+			ISRCopyBuffer.pBufferBase, size, ISRCopyBuffer.u4BufferSize);
 
 	size = Align64ByteSize(size);
